requirements.txt
streamlit
pandas
plotly
numpy
supabase
python-dotenv